more ints

+++ main.cpp
@@ -168,13 +168,13 @@ void draw_globe(uint8_t *framebuffer) {
 	const uint8_t  *globdata = GLOBDATA_BIN;
 	const uint8_t  *map      = MAP_BIN;
 
-	uint16_t cs_1CA6 = 1;    // offset into globdata
-	int16_t cs_1CB4 = -FRAMEBUFFER_WIDTH; // screen width
+	int cs_1CA6 = 1;    // offset into globdata
+	int cs_1CB4 = -FRAMEBUFFER_WIDTH; // screen width
 
-	const uint16_t globe_center_xy_offset1 = frame_buffer_offset(160, 79);
-	uint16_t cs_1CB0 = globe_center_xy_offset1;
-	uint16_t cs_1CB2 = globe_center_xy_offset1;
-	uint16_t cs_1CAE = globe_center_xy_offset1 - 1;
+	const int globe_center_xy_offset1 = frame_buffer_offset(160, 79);
+	int cs_1CB0 = globe_center_xy_offset1;
+	int cs_1CB2 = globe_center_xy_offset1;
+	int cs_1CAE = globe_center_xy_offset1 - 1;
 
 	bool drawing_southern_hemisphere = false;
 
@@ -215,10 +215,10 @@ void draw_globe(uint8_t *framebuffer) {
 				uint16_t grlt_2{};
 			};
 
-			auto func1 = [](const uint8_t* globdata_, const globe_rotation_lookup_table_t& rotation_lookup_table, const int16_t ofs1, const int16_t base_ofs) {
+			auto func1 = [](const uint8_t* globdata_, const globe_rotation_lookup_table_t& rotation_lookup_table, const int16_t ofs1, const int base_ofs) {
 				const int8_t lo_ofs1 = lo(ofs1);
 
-				const int8_t offset1 = ( lo_ofs1 < 0 ) ? -lo_ofs1 : lo_ofs1;
+				const int offset1 = ( lo_ofs1 < 0 ) ? -lo_ofs1 : lo_ofs1;
 				const uint8_t* sub_globdata = &globdata_[base_ofs + offset1];
 
 				// sub_globdata contains sizeof(uint16_t)-offsets to the entry but we need a logical index to our entry wich is 4*uint16_t
@@ -241,7 +241,7 @@ void draw_globe(uint8_t *framebuffer) {
 					value += adjust1;
 				}
 				value += adjust2;
-				return value;
+				return int(value);
 			};
 
 			auto pixel_color = [](uint8_t value) {
